Fixed: project was not working in terminal outside pycharm
