+++ freelancer/static/css/bootstrap-icons.css
