+++ freelancer/static/css/fonts/bootstrap-icons.woff2
